#407: Moved PassiveEquipmentCategoryElementsResistanceParser into the model folder
For now it is the same file.
