+++ src/app/ffbe/mappers/effects/passive-effect-parser.factory.ts
@@ -36,7 +36,7 @@ import {PassiveGilsRateParser} from './passives/passive-gils-rate.parser';
 import {PassiveExperienceRateParser} from './passives/passive-experience-rate.parser';
 import {PassiveCombatRateDecreaseParser} from './passives/passive-combat-rate-decrease.parser';
 import {PassiveSkillAliveAllyActivationParser} from './passives/passive-skill-alive-ally-activation.parser';
-import {PassiveEquipmentCategoryElementsResistanceParser} from './passives/passive-equipment-category-elements-resistance.parser';
+import {PassiveEquipmentCategoryElementsResistanceParser} from '../../model/effects/passives/passive-equipment-category-elements-resistance.parser';
 import {PassiveItemsHealingPotencyIncreaseParser} from './passives/passive-items-healing-potency-increase.parser';
 import {PassiveExplorationStepRegenParser} from './passives/passive-exploration-step-regen.parser';
 import {PassiveGilsWhileStealingParser} from './passives/passive-gils-while-stealing.parser';
